whisper-server/requirements.txt
openai-whisper>=20231117
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0

